Time: 330 ms (5.65%), Space: 39.4 MB (83.92%) - LeetHub

+++ 400-nth-digit/400-nth-digit.java
@@ -4,7 +4,7 @@ class Solution {
         if(n==Integer.MAX_VALUE) return 2; 
         int weight = 1;
         int counter = 1;
-        long place =10;
+        int place =10;
         int num = 1;
         while(counter+weight <=n){
             counter+=weight;
